+++ tests/test_types_tensor_subclasses.py
@@ -34,6 +34,7 @@ from torchwrench.types.tensor_subclasses import (
     ShortTensor1D,
     SignedIntegerTensor,
     SignedIntegerTensor3D,
+    Tensor,
     Tensor0D,
     Tensor1D,
     UnsignedIntegerTensor,
@@ -221,6 +222,7 @@ class TestTensorTyping(TestCase):
             tensor = tensor_cls()
 
             assert isinstance(tensor, torch.Tensor)
+            assert isinstance(tensor, Tensor)
 
         warnings.resetwarnings()